Proper secp256k1 parsing
